Clean up code after moving admin to rpc
* Move grpchelper package to rpc package
* Rename authInterceptor to adminAuthInterceptor
* Simply condition block with an early return pattern
